+++ src/main/java/frc/util/combo/LimitDeadzoneCombo.java
@@ -1,10 +1,17 @@
-package frc.util;
+package frc.util.combo;
+
+import frc.util.*;;
+
+
+//Requires TriggerDeadzone.java and MotorLimits.java
 
 public class LimitDeadzoneCombo{
+
     public static double setDeadzoneLimits(double controlInput, double deadzoneValue){
+        
         double output = controlInput;
-        output = MotorLimits.getMotorLimits(output);
         output = TriggerDeadzone.getDeadzoneResult(output, deadzoneValue);
+        output = MotorLimits.getMotorLimits(output);
         return output;
